requirements.txt
PyQt5==5.15.11
pyautogui==0.9.54
Pillow==11.1.0
opencv-python>=4.5.0 